Update another sanitizer test for C++14.
Follow-up to r320251.

llvm-svn: 320284

+++ compiler-rt/test/asan/TestCases/Posix/halt_on_error-torture.cc
@@ -26,7 +26,7 @@ size_t niter = 10;
 
 void random_delay(unsigned *seed) {
   *seed = 1664525 * *seed + 1013904223;
-  struct timespec delay = { 0, (*seed % 1000) * 1000 };
+  struct timespec delay = { 0, static_cast<long>((*seed % 1000) * 1000) };
   nanosleep(&delay, 0);
 }
